HomeWork to Lesson DD9.11.23
Sorry for delay - the deadline was on 19th November, 23, but I had many problems. Now it looks much better, but I have some questions any more: 1. How to insert in input emails? How to insert Passwords? For the both the input type to be innerText or Numbers? Also, I didn't finish with style of check-box. I try to complete it to the end of the week.